+++ devicelayer/include/deviceLayer/DeviceLayerMock.h
@@ -59,6 +59,7 @@ public:
   MOCK_METHOD2(updateFirmwareImage, int(int, std::vector<unsigned char>&));
   MOCK_CONST_METHOD0(getFreeCmaMemory, size_t());
   MOCK_CONST_METHOD1(getDmaInfo, DmaInfo(int));
+  MOCK_METHOD3(reinitDeviceInstance, void(int device, bool masterMinionOnly, std::chrono::milliseconds timeout));
 
   void Delegate() {
     ON_CALL(*this, sendCommandMasterMinion)
@@ -139,6 +140,10 @@ public:
     ON_CALL(*this, clearDeviceAttributes).WillByDefault([this](int device, std::string relGroupPath) {
       delegate_->clearDeviceAttributes(device, relGroupPath);
     });
+    ON_CALL(*this, reinitDeviceInstance)
+      .WillByDefault([this](int device, bool masterMinionOnly, std::chrono::milliseconds timeout) {
+        delegate_->reinitDeviceInstance(device, masterMinionOnly, timeout);
+      });
   }
 
 private: