+++ devicelayer/src/DeviceSysEmuMulti.h
@@ -51,6 +51,7 @@ public:
   DmaInfo getDmaInfo(int device) const override;
   std::string getDeviceAttribute(int device, std::string relAttrPath) const override;
   void clearDeviceAttributes(int device, std::string relGroupPath) const override;
+  void reinitDeviceInstance(int device, bool masterMinionOnly, std::chrono::milliseconds timeout) override;
 
 private:
   DeviceSysEmu& getDevice(int device);