+++ devicelayer/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- [SW-14409] Added new method reinitDeviceInstance() which facilitates single device reinitialization.
 ### Changed
 ### Deprecated
 ### Removed